+++ packages/headless/src/controllers/knowledge/generated-answer/headless-answerapi-generated-answer.ts
@@ -15,6 +15,7 @@ import {SearchEngine} from '../../../app/search-engine/search-engine.js';
 import {
   resetAnswer,
   sendGeneratedAnswerFeedback,
+  setCannotAnswer,
   updateAnswerConfigurationId,
 } from '../../../features/generated-answer/generated-answer-actions.js';
 import {GeneratedAnswerFeedback} from '../../../features/generated-answer/generated-answer-analytics-actions.js';
@@ -99,6 +100,10 @@ const subscribeToSearchRequest = (
 
     if (!lastTriggerParams || triggerParams.q.length === 0) {
       lastTriggerParams = triggerParams;
+    }
+
+    if (triggerParams.q.length === 0 && !!state.generatedAnswer.cannotAnswer) {
+      engine.dispatch(setCannotAnswer(false));
     }
 
     if (